added api
i added the pokemon api with a fetch and used a try and catch to catch a problem, also added a router with homepage and pokemon page.
pokemon page is first gen of pokemon in a list were the names of the pokemon are clickable and clicking brings you to the pokemons page where you can see the pokemon sprite, weight and heigth. added search balk  (can search for pokemon names)